requirements.txt
Django==1.11.4
gunicorn==19.6.0
django-webpack-loader==0.5.0
djangorestframework==3.6.4
psycopg2==2.7.3
pyshark==0.3.7.5
pyshark-parser==0.1

